调整优化代码等;增加插件常用功能;文档更新;

+++ models/global.go
@@ -4,12 +4,12 @@ import "github.com/dreamlu/gt/tool/type/time"
 
 // 通用模型
 type ModelCom struct {
-	ID         uint       `gorm:"type:bigint(20) AUTO_INCREMENT;PRIMARY_KEY;" json:"id"`
+	ID         uint64     `gorm:"type:bigint(20) AUTO_INCREMENT;PRIMARY_KEY;" json:"id"`
 	Createtime time.CTime `gorm:"type:datetime;DEFAULT:CURRENT_TIMESTAMP" json:"createtime"` // 创建时间自动生成
 }
 
 // 账号关联
 type AdminCom struct {
 	ModelCom
-	AdminID uint `json:"admin_id" gorm:"type:bigint(20);INDEX:查询索引admin_id"`
+	AdminID uint64 `json:"admin_id" gorm:"type:bigint(20);INDEX:查询索引admin_id"`
 }
